requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
et_xmlfile==2.0.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
openpyxl==3.1.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
